pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "gpumanager"
version = "0.1.0"
description = "Lightweight NVIDIA GPU utilization sampler, Slack reporter, and systemd timer helper"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [
  { name = "OpenAI Codex" }
]
keywords = ["gpu", "nvidia", "slack", "systemd", "monitoring", "cli"]
dependencies = [
  "tomli>=1.1.0; python_version < '3.11'",
  "backports.zoneinfo>=0.2.1; python_version < '3.9'",
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: Science/Research",
  "Intended Audience :: System Administrators",
  "License :: OSI Approved :: MIT License",
  "Operating System :: POSIX :: Linux",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: System :: Monitoring",
  "Topic :: System :: Systems Administration",
]

[project.scripts]
gpumanager = "gpumanager.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
